smooth interpolation for the following stuff: ghoul2 animation, wave sin for shaders, entities in cgame; fixed frozen effects in local server, fixed wave sawtooth interpolation

+++ codemp/rd-vanilla/tr_local.h
@@ -1319,7 +1319,7 @@ Ghoul2 Insert End
 */
 //====================================================================
 
-float R_NoiseGet4f( float x, float y, float z, float t );
+float R_NoiseGet4f( float x, float y, float z, double t );
 void  R_NoiseInit( void );
 
 void R_SwapBuffers( int );
