+++ codemp/rd-vanilla/tr_noise.cpp
@@ -13,7 +13,7 @@
 static float s_noise_table[NOISE_SIZE];
 static int s_noise_perm[NOISE_SIZE];
 
-#define LERP( a, b, w ) ( a * ( 1.0f - w ) + b * w )
+#define LERP( a, b, w ) ( a * ( 1.0 - w ) + b * w )
 
 static float GetNoiseValue( int x, int y, int z, int t )
 {
@@ -42,11 +42,11 @@ void R_NoiseInit( void )
 	}
 }
 
-float R_NoiseGet4f( float x, float y, float z, float t )
+float R_NoiseGet4f( float x, float y, float z, double t )
 {
 	int i;
 	int ix, iy, iz, it;
-	float fx, fy, fz, ft;
+	double fx, fy, fz, ft;
 	float front[4];
 	float back[4];
 	float fvalue, bvalue, value[2], finalvalue;
